src/tools/stt_engine/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
openai-whisper>=20230918
